+++ contrib/opencensus-ext-azure/opencensus/ext/azure/common/transport.py
@@ -129,16 +129,18 @@ class TransportMixin(object):
                     duration = _requests_map.get('duration', 0)
                     _requests_map['duration'] = duration + (end_time - start_time)  # noqa: E501
             if exception is not None:
-                if self._is_stats_exporter() and not state.get_statsbeat_initial_success():
-                    # If ingestion threshold during statsbeat initialization is reached, return back code to shut it down
-                    if _statsbeat_failed_to_ingest():
-                        return -2
                 if self._check_stats_collection():
                     with _requests_lock:
                         if exception >= 0:
                             _requests_map['retry'] = _requests_map.get('retry', 0) + 1  # noqa: E501
                         else:
                             _requests_map['exception'] = _requests_map.get('exception', 0) + 1  # noqa: E501
+                if self._is_stats_exporter() and \
+                    not state.get_statsbeat_shutdown() and \
+                        not state.get_statsbeat_initial_success():
+                    # If ingestion threshold during statsbeat initialization is reached, return back code to shut it down
+                    if _statsbeat_failed_to_ingest():
+                        return -2
                 return exception
 
         text = 'N/A'
@@ -154,7 +156,9 @@ class TransportMixin(object):
             except Exception:
                 pass
 
-        if self._is_stats_exporter() and not state.get_statsbeat_initial_success():
+        if self._is_stats_exporter() and \
+            not state.get_statsbeat_shutdown() and \
+                not state.get_statsbeat_initial_success():
             # If statsbeat exporter, record initialization as success if appropriate status code is returned
             if _reached_ingestion_status_code(response.status_code):
                 state.set_statsbeat_initial_success(True)
@@ -298,5 +302,5 @@ def _reached_ingestion_status_code(status_code):
 
 def _statsbeat_failed_to_ingest():
     # increment failure counter for sending statsbeat if still in initialization
-    state.increment_statsbeats_initial_failure_count()
+    state.increment_statsbeat_initial_failure_count()
     return state.get_statsbeat_initial_failure_count() >= 3