+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Statsbeat - shutdown stats when hit error/exception threshold
+([#1113](https://github.com/census-instrumentation/opencensus-python/pull/1113))
+
 ## 1.1.4
 Released 2022-04-20